requirements.txt
pymongo
pycrypto